+++ app/src/main/java/com/example/cookshop/view/main/FragmentSyc.java
@@ -1,0 +1,36 @@
+package com.example.cookshop.view.main;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+import com.example.cookshop.R;
+
+
+public class FragmentSyc extends Fragment
+{
+
+View view;
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        view =  inflater.inflate(R.layout.fragment_sync, container, false);
+
+        return view;
+    }
+
+
+
+
+}
