+++ app/src/main/java/com/example/cookshop/view/main/MainActivity.java
@@ -4,6 +4,8 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.MenuItem;
+import android.view.View;
+import android.widget.Toast;
 
 import com.example.cookshop.R;
 import com.example.cookshop.model.database.DatabaseHelper;
@@ -11,7 +13,7 @@ import com.example.cookshop.model.listManagement.AvailableListService;
 import com.example.cookshop.model.listManagement.DataAccess;
 import com.example.cookshop.model.listManagement.RecipeListService;
 import com.example.cookshop.model.listManagement.ShoppingListService;
-import com.example.cookshop.view.articleViewUpdateAdd.ArticleViewer;
+import com.example.cookshop.view.SynchronizeActivity;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 import androidx.annotation.NonNull;
@@ -97,9 +99,9 @@ public class MainActivity extends AppCompatActivity
                     Log.d(TAG, "onNavigationItemSelected: replacing current fragment with fragment toCook");
                     selectedFragment = FragmentShoppingList.class;
                     break;
-                case R.id.nav_recommended:
-                    Log.d(TAG, "onNavigationItemSelected: replacing current fragment with fragment recommended");
-                    selectedFragment = FragmentRecommendedList.class;
+                case R.id.nav_sync:
+                    Log.d(TAG, "onNavigationItemSelected: replacing current fragment with fragment sync");
+                    selectedFragment = FragmentSyc.class;
                     break;
                 case R.id.nav_recipes:
                     Log.d(TAG, "onNavigationItemSelected: replacing current fragment with fragment recipes");
@@ -135,4 +137,16 @@ public class MainActivity extends AppCompatActivity
 
 
 
+    public void shoppingListSyncOnClick(View view)
+    {
+        Intent syncIntent = new Intent(getApplicationContext(), SynchronizeActivity.class);
+        startActivity(syncIntent);
+    }
+
+    public void recipeListSyncOnClick(View view)
+    {
+        Toast toast=Toast.makeText(getApplicationContext(),"Not available at the moment",Toast.LENGTH_SHORT);
+        toast.setMargin(50,50);
+        toast.show();
+    }
 }
